requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic[email]>=2.9.0
email-validator>=2.2.0
python-multipart>=0.0.12
bcrypt>=4.2.0
PyJWT>=2.9.0